+++ config/ipfs_test.go
@@ -1,0 +1,7 @@
+package config
+
+import "testing"
+
+func TestDefaultIPFSConfig(t *testing.T) {
+	assertStorageConfigSerialisation(t, DefaultIPFSConfig())
+}