+++ config/storage/cloud_test.go
@@ -1,9 +1,0 @@
-package storage
-
-import "testing"
-
-func TestDefaultCloudConfig(t *testing.T) {
-	assertStorageConfigSerialisation(t, DefaultCloudConfig("aws"))
-	assertStorageConfigSerialisation(t, DefaultCloudConfig("azure"))
-	assertStorageConfigSerialisation(t, DefaultCloudConfig("gcp"))
-}