+++ config/storage/ipfs_test.go
@@ -1,7 +1,0 @@
-package storage
-
-import "testing"
-
-func TestDefaultIPFSConfig(t *testing.T) {
-	assertStorageConfigSerialisation(t, DefaultIPFSConfig())
-}