+++ config/storage/memory_test.go
@@ -1,7 +1,0 @@
-package storage
-
-import "testing"
-
-func TestDefaultMemoryConfig(t *testing.T) {
-	assertStorageConfigSerialisation(t, DefaultMemoryConfig())
-}