+++ grant/grant_test.go
@@ -4,7 +4,7 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/monax/hoard/v4/config/secrets"
+	"github.com/monax/hoard/v4/config"
 	"github.com/monax/hoard/v4/reference"
 	"github.com/stretchr/testify/assert"
 )
@@ -15,12 +15,12 @@ func TestGrants(t *testing.T) {
 	keyPrivate, err := ioutil.ReadFile("private.key.asc")
 	assert.NoError(t, err)
 
-	testPGP := secrets.OpenPGPSecret{
+	testPGP := config.OpenPGPSecret{
 		PrivateID: "10449759736975846181",
 		Data:      keyPrivate,
 	}
 
-	testSecrets := secrets.Manager{
+	testSecrets := config.SecretsManager{
 		Provider: func(_ string) ([]byte, error) {
 			return nil, nil
 		},