+++ project/history.go
@@ -33,7 +33,7 @@ var History relic.ImmutableHistory = relic.NewHistory("Monax Hoard", "https://gi
 		``,
 		"4.0.0 - 2019-05-21",
 		`###Fixed
-- [BUILD] Change hoard.pb.go to services/services.pb.go
+- [BUILD] Change hoard.pb.go to services/api.pb.go
 `,
 		"3.2.1 - 2019-04-24",
 		`### Fixed