+++ storage/memory_test.go
@@ -1,7 +1,0 @@
-package storage
-
-import "testing"
-
-func TestMemoryStore(t *testing.T) {
-	testStore(t, NewMemoryStore())
-}