+++ stores/memory_test.go
@@ -1,0 +1,9 @@
+package stores
+
+import (
+	"testing"
+)
+
+func TestMemoryStore(t *testing.T) {
+	RunTests(t, NewMemoryStore())
+}